# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users (admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get a user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update a user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete a user (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/reviews": {
      "get": {
        "operationId": "ReviewsController_findAll",
        "summary": "Get all reviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "ReviewsController_create",
        "summary": "Create a new review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input or session not completed"
          },
          "404": {
            "description": "Consultant, client, or session not found"
          },
          "409": {
            "description": "Session already has a review"
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/reviews/consultant/{consultantId}": {
      "get": {
        "operationId": "ReviewsController_findByConsultant",
        "summary": "Get reviews by consultant ID",
        "parameters": [
          {
            "name": "consultantId",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns reviews for the specified consultant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/reviews/{id}": {
      "get": {
        "operationId": "ReviewsController_findOne",
        "summary": "Get a review by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "404": {
            "description": "Review not found"
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ReviewsController_remove",
        "summary": "Delete a review (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/messaging/send": {
      "post": {
        "operationId": "MessagingController_sendMessage",
        "summary": "Send a message to a client",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "messageId": {
                      "type": "string",
                      "example": "msg_12345678"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input"
          },
          "403": {
            "description": "Forbidden - Requires consultant or admin role"
          },
          "404": {
            "description": "Messenger type not found"
          }
        },
        "tags": [
          "messaging"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Family Counseling API",
    "description": "\n      # Family Counseling API Documentation\n      \n      This API provides the backend services for the Family Counseling Platform.\n\n      ## Authorization\n      Most endpoints require JWT authentication. After logging in, include the JWT token in the Authorization header:\n      ```\n      Authorization: Bearer YOUR_JWT_TOKEN\n      ```\n      \n      ## Role-based access\n      Different endpoints have different role requirements. The API supports these roles:\n      - admin: Full access to all endpoints\n      - consultant: Access to consultant-specific features\n      - client: Access to client-specific features\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "fullName": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "consultant",
              "client"
            ],
            "default": "client"
          },
          "phoneNumber": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "fullName",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "phoneNumber": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          }
        }
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Status of the operation",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation successful"
          },
          "data": {
            "type": "object",
            "description": "Response data",
            "example": {}
          },
          "error": {
            "type": "string",
            "description": "Error description if any",
            "example": null
          }
        },
        "required": [
          "success",
          "message",
          "data",
          "error"
        ]
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Review comment"
          }
        },
        "required": [
          "rating",
          "comment"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "consultantId": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Review comment"
          }
        },
        "required": [
          "consultantId",
          "clientId",
          "sessionId",
          "rating",
          "comment"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "messengerType": {
            "type": "string",
            "enum": [
              "telegram",
              "whatsapp"
            ]
          },
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "recipientId",
          "text",
          "messengerType"
        ]
      },
...
    }
  }
}